+++ R/res2tbl.R
@@ -2,10 +2,12 @@
 #'
 #' Generate a tidy table with the results of DESeq
 #'
-#' @param deseqresult A \code{\link{DESeqResults}} object
+#' @param deseqresult A [DESeqResults()] object
 #'
 #' @return A "tidy" data.frame with all genes
 #' @export
+#' 
+#' @importFrom mosdef deresult_to_df
 #'
 #' @examples
 #'
@@ -16,15 +18,16 @@
 #' res <- DESeq2::results(dds)
 #' deseqresult2tbl(res)
 deseqresult2tbl <- function(deseqresult) {
-  # library("dplyr")
-  if (!is(deseqresult, "DESeqResults")) stop("Not a DESeqResults object.")
-  deseqresult <- as.data.frame(deseqresult)
-
-  deseqresult <- cbind(rownames(deseqresult), deseqresult)
-  names(deseqresult)[1] <- "id"
-  deseqresult$id <- as.character(deseqresult$id)
-
-  dplyr::arrange(deseqresult, .data$padj)
+  .Deprecated(old = "deseqresult2tbl", new = "mosdef::deresult_to_df", 
+              msg = paste0(
+                "Please use `mosdef::deresult_to_df()` in replacement of the `deseqresult2tbl()` function, ",
+                "originally located in the ideal package. \nCheck the manual page for ",
+                "`?mosdef::deresult_to_df()` to see the details on how to use it, e.g. ",
+                "refer to the new parameter definition and naming"))
+  
+  res_de <- mosdef::deresult_to_df(deseqresult)
+  
+  return(res_de)
 }
 
 
@@ -34,11 +37,13 @@ deseqresult2tbl <- function(deseqresult) {
 #'
 #' Generate a tidy table with the DE genes from the results of DESeq
 #'
-#' @param deseqresult A \code{\link{DESeqResults}} object
+#' @param deseqresult A [DESeqResults()] object
 #' @param FDR Numeric value, the significance level for thresholding adjusted p-values
 #'
 #' @return A "tidy" data.frame with only genes marked as differentially expressed
 #' @export
+#' 
+#' @importFrom mosdef deresult_to_df
 #'
 #' @examples
 #'
@@ -50,22 +55,14 @@ deseqresult2tbl <- function(deseqresult) {
 #' deseqresult2DEgenes(res)
 deseqresult2DEgenes <- function(deseqresult,
                                 FDR = 0.05) {
-  # library("dplyr")
-  if (!is(deseqresult, "DESeqResults")) stop("Not a DESeqResults object.")
-  deseqresult <- as.data.frame(deseqresult)
-
-  deseqresult <- cbind(rownames(deseqresult), deseqresult)
-  names(deseqresult)[1] <- "id"
-  deseqresult$id <- as.character(deseqresult$id)
-
-  # deseqresult$id <- rownames(deseqresult)
-  # rownames(deseqresult) <- NULL
-  # deseqresult <- dplyr::tbl_df(deseqresult)
-  # if("symbol" %in% names(deseqresult))
-  #   deseqresult <- dplyr::select(deseqresult, id, baseMean, log2FoldChange:symbol)
-  # else
-  #   deseqresult <- dplyr::select(deseqresult, id, baseMean, log2FoldChange:padj)
-  tmp <- dplyr::arrange(deseqresult, .data$padj)
-  res <- tmp[!(is.na(tmp$padj)) & tmp$padj <= FDR, ]
-  res
+  .Deprecated(old = "deseqresult2DEgenes", new = "mosdef::deresult_to_df", 
+              msg = paste0(
+                "Please use `mosdef::deresult_to_df()` in replacement of the `deseqresult2DEgenes()` function, ",
+                "originally located in the ideal package. \nCheck the manual page for ",
+                "`?mosdef::deresult_to_df()` to see the details on how to use it, e.g. ",
+                "refer to the new parameter definition and naming"))
+  
+  res_de <- mosdef::deresult_to_df(deseqresult, FDR = FDR)
+  
+  return(res_de)
 }